Permissions improvements
- Swap classes for enums and a lookup
- Add helper methods for mutual exclusivity and scope validation
- Adjust tests
