--- src/SendGrid/Permissions/Scopes/Credentials.cs
@@ -1,24 +1,0 @@
-﻿namespace SendGrid.Permissions.Scopes
-{
-    /// <summary>
-    /// Scopes for credentials
-    /// </summary>
-    /// <seealso cref="SendGridPermissionScope" />
-    public class Credentials : SendGridPermissionScope
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="Credentials"/> class.
-        /// </summary>
-        public Credentials()
-            : base("Credentials")
-        {
-            this.Scopes = new[]
-            {
-                "credentials.create",
-                "credentials.delete",
-                "credentials.read",
-                "credentials.update"
-            };
-        }
-    }
-}
