+++ src/SendGrid/Permissions/SendGridPermission.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace SendGrid.Permissions
+{
+#pragma warning disable CS1591 // Missing XML comment for publicly visible type or member
+    public enum SendGridPermission
+
+    {
+        Admin,
+        Alerts,
+        ApiKeys,
+        ASMGroups,
+        Billing,
+        Categories,
+        Clients,
+        Credentials,
+        DomainAuthentication,
+        IPs,
+        Mail,
+        MailSettings,
+        MarketingCampaigns,
+        Newsletter,
+        PartnerSettings,
+        ReverseDNS,
+        ScheduledSends,
+        Stats,
+        Subusers,
+        Suppressions,
+        Teammates,
+        Templates,
+        Tracking,
+        UserSettings,
+        Webhook
+    }
+#pragma warning restore CS1591 // Missing XML comment for publicly visible type or member
+}
