+++ src/SendGrid/SendGrid.csproj
@@ -40,6 +40,10 @@
   </ItemGroup>
 
   <ItemGroup>
+    <None Include="..\..\.editorconfig" Link=".editorconfig" />
+  </ItemGroup>
+
+  <ItemGroup>
     <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0" PrivateAssets="All" Condition=" '$(OS)' != 'Windows_NT' " />
     <PackageReference Include="Newtonsoft.Json" Version="9.0.1" />
     <PackageReference Include="starkbank-ecdsa" Version="[1.2.0, 2.0.0)" />
