+++ tests/SendGrid.Tests/SendGridClientExtensionsTests.cs
@@ -7,7 +7,6 @@ using Moq;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using SendGrid.Permissions;
-using SendGrid.Permissions.Scopes;
 using Xunit;
 
 namespace SendGrid.Tests
@@ -48,7 +47,7 @@ namespace SendGrid.Tests
                 .ReturnsAsync(new Response(HttpStatusCode.Created, null, null));
 
             var builder = new SendGridPermissionsBuilder();
-            builder.AddPermissionsFor<Alerts>();
+            builder.AddPermissionsFor(SendGridPermission.Alerts);
             var response = await mockClient.Object.CreateApiKey(builder, "Alerts Test");
 
             Assert.Equal(4, requestedScopes.Length);
@@ -58,42 +57,6 @@ namespace SendGrid.Tests
             Assert.Contains("alerts.update", requestedScopes);
 
             Assert.Equal("Alerts Test", requestedApiKeyName);
-        }
-
-        [Fact]
-        public async Task MasksPermissionsForAGivenClient_Realsie()
-        {
-            // create api key with only alerts.read
-            var client = new SendGridClient("");
-
-            var builder = await client.CreateMaskedPermissionsBuilderForClient();
-
-            builder.Include("mail.send", "alerts.read");
-            var scopes = builder.Build().ToArray();
-
-            Assert.Single(scopes);
-            Assert.Contains("alerts.read", scopes);
-        }
-
-        [Fact]
-        public async Task CallsApiWithBuiltScopes_Realsie()
-        {
-            var client = new SendGridClient("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");
-            var builder = new SendGridPermissionsBuilder();
-            builder.AddPermissionsFor<Alerts>();
-            var response = await client.CreateApiKey(builder, "Alerts Test");
-            var json = await response.DeserializeResponseBodyAsync(response.Body);
-
-            var scopes = (json["scopes"] as JArray).Select(x => x.Value<string>()).ToArray();
-            var name = json["name"];
-
-            Assert.Equal(4, scopes.Length);
-            Assert.Contains("alerts.create", scopes);
-            Assert.Contains("alerts.delete", scopes);
-            Assert.Contains("alerts.read", scopes);
-            Assert.Contains("alerts.update", scopes);
-
-            Assert.Equal("Alerts Test", name);
         }
     }
 }
